make main method cleaner
