+++ validator/BUILD.bazel
@@ -21,17 +21,13 @@ go_library(
         "//shared/interop:go_default_library",
         "//shared/keystore:go_default_library",
         "//shared/logutil:go_default_library",
-        "//shared/params:go_default_library",
         "//shared/version:go_default_library",
-        "//validator/accounts:go_default_library",
         "//validator/flags:go_default_library",
         "//validator/node:go_default_library",
         "@com_github_joonix_log//:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
         "@com_github_x_cray_logrus_prefixed_formatter//:go_default_library",
-        "@org_golang_x_crypto//ssh/terminal:go_default_library",
         "@org_uber_go_automaxprocs//:go_default_library",
     ],
 )
