+++ validator/accounts/BUILD.bazel
@@ -2,15 +2,22 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["account.go"],
+    srcs = [
+        "account.go",
+        "interop.go",
+        "keys.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/validator/accounts",
     visibility = ["//validator:__subpackages__"],
     deps = [
         "//contracts/deposit-contract:go_default_library",
+        "//shared/bls:go_default_library",
+        "//shared/interop:go_default_library",
         "//shared/keystore:go_default_library",
         "//shared/params:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
+        "@org_uber_go_automaxprocs//:go_default_library",
     ],
 )
