+++ validator/node/BUILD.bazel
@@ -15,7 +15,10 @@ go_test(
 
 go_library(
     name = "go_default_library",
-    srcs = ["node.go"],
+    srcs = [
+        "keyutils.go",
+        "node.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/validator/node",
     visibility = ["//validator:__subpackages__"],
     deps = [
@@ -28,10 +31,12 @@ go_library(
         "//shared/prometheus:go_default_library",
         "//shared/tracing:go_default_library",
         "//shared/version:go_default_library",
+        "//validator/accounts:go_default_library",
         "//validator/client:go_default_library",
         "//validator/flags:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
+        "@org_golang_x_crypto//ssh/terminal:go_default_library",
     ],
 )
